+++ packages/extension-koni-base/src/api/tokens/evm/defaultEvmToken.ts
@@ -331,6 +331,14 @@ export const DEFAULT_EVM_TOKENS: EvmTokenJson = {
       decimals: 18,
       chain: 'ethereum',
       type: CustomTokenType.erc20
+    },
+    {
+      name: 'TFA',
+      symbol: 'TFA',
+      smartContract: '0xE065ffaf3f7dED69BB5cf5FDd1Fd1dDA2EEe8493',
+      decimals: 18,
+      chain: 'moonbeam',
+      type: CustomTokenType.erc20
     }
   ],
   erc721: [