+++ packages/extension-koni-base/src/background/handlers/State.ts
@@ -228,7 +228,7 @@ export default class KoniState extends State {
 
             mergedNetworkMap[key].crowdloanUrl = storedNetwork.crowdloanUrl;
             mergedNetworkMap[key].blockExplorer = storedNetwork.blockExplorer;
-            mergedNetworkMap[key].currentProviderMode = mergedNetworkMap[key].currentProvider.startsWith('http') ? 'http' : 'ws';
+            mergedNetworkMap[key].currentProviderMode = (mergedNetworkMap[key].currentProvider || '').startsWith('http') ? 'http' : 'ws';
           } else {
             if (Object.keys(PREDEFINED_GENESIS_HASHES).includes(storedNetwork.genesisHash)) { // merge networks with same genesis hash
               // @ts-ignore
@@ -254,11 +254,17 @@ export default class KoniState extends State {
       }
 
       for (const [key, network] of Object.entries(this.networkMap)) {
+        const currentProvider = getCurrentProvider(network);
+
+        if (!currentProvider) {
+          continue;
+        }
+
         if (network.active) {
-          this.apiMap.dotSama[key] = initApi(key, getCurrentProvider(network), network.isEthereum);
+          this.apiMap.dotSama[key] = initApi(key, currentProvider, network.isEthereum);
 
           if (network.isEthereum && network.isEthereum) {
-            this.apiMap.web3[key] = initWeb3Api(getCurrentProvider(network));
+            this.apiMap.web3[key] = initWeb3Api(currentProvider);
           }
         }
       }
@@ -795,6 +801,11 @@ export default class KoniState extends State {
 
   public setHistory (address: string, network: string, item: TransactionHistoryItemType | TransactionHistoryItemType[], callback?: (items: TransactionHistoryItemType[]) => void): void {
     let items: TransactionHistoryItemType[];
+    const networkInfo = this.getNetworkMap()[network];
+
+    if (!networkInfo) {
+      return;
+    }
 
     if (item && !Array.isArray(item)) {
       item.origin = 'app';
@@ -802,6 +813,14 @@ export default class KoniState extends State {
     } else {
       items = item;
     }
+
+    items.forEach((item) => {
+      item.feeSymbol = networkInfo.nativeToken;
+
+      if (!item.changeSymbol) {
+        item.changeSymbol = networkInfo.nativeToken;
+      }
+    });
 
     if (items.length) {
       this.getAccountAddress().then((currentAddress) => {
@@ -990,6 +1009,10 @@ export default class KoniState extends State {
     }
 
     return [checkingAddress];
+  }
+
+  public getAllAddresses (): string[] {
+    return Object.keys(accounts.subject.value);
   }
 
   public getBalance (reset?: boolean): BalanceJson {
@@ -1382,7 +1405,7 @@ export default class KoniState extends State {
         this.networkMap[data.key].customProviders = data.customProviders;
       }
 
-      if (data.currentProvider !== this.networkMap[data.key].currentProvider) {
+      if (data.currentProvider !== this.networkMap[data.key].currentProvider && data.currentProvider) {
         this.networkMap[data.key].currentProvider = data.currentProvider;
         this.networkMap[data.key].currentProviderMode = data.currentProvider.startsWith('ws') ? 'ws' : 'http';
       }
@@ -1419,10 +1442,14 @@ export default class KoniState extends State {
         delete this.apiMap.web3[data.key];
       }
 
-      this.apiMap.dotSama[data.key] = initApi(data.key, getCurrentProvider(data), data.isEthereum);
+      const currentProvider = getCurrentProvider(data);
 
-      if (data.isEthereum && data.isEthereum) {
-        this.apiMap.web3[data.key] = initWeb3Api(getCurrentProvider(data));
+      if (currentProvider) {
+        this.apiMap.dotSama[data.key] = initApi(data.key, currentProvider, data.isEthereum);
+
+        if (data.isEthereum && data.isEthereum) {
+          this.apiMap.web3[data.key] = initWeb3Api(currentProvider);
+        }
       }
     }
 
@@ -1559,10 +1586,14 @@ export default class KoniState extends State {
     const networkData = this.networkMap[networkKey];
 
     this.lockNetworkMap = true;
-    this.apiMap.dotSama[networkKey] = initApi(networkKey, getCurrentProvider(networkData), networkData.isEthereum);
+    const currentProvider = getCurrentProvider(networkData);
 
-    if (networkData.isEthereum && networkData.isEthereum) {
-      this.apiMap.web3[networkKey] = initWeb3Api(getCurrentProvider(networkData));
+    if (currentProvider) {
+      this.apiMap.dotSama[networkKey] = initApi(networkKey, currentProvider, networkData.isEthereum);
+
+      if (networkData.isEthereum && networkData.isEthereum) {
+        this.apiMap.web3[networkKey] = initWeb3Api(currentProvider);
+      }
     }
 
     networkData.active = true;
@@ -1601,10 +1632,14 @@ export default class KoniState extends State {
     this.networkMapStore.set('NetworkMap', this.networkMap);
 
     for (const key of targetNetworkKeys) {
-      this.apiMap.dotSama[key] = initApi(key, getCurrentProvider(this.networkMap[key]), this.networkMap[key].isEthereum);
+      const currentProvider = getCurrentProvider(this.networkMap[key]);
 
-      if (this.networkMap[key].isEthereum && this.networkMap[key].isEthereum) {
-        this.apiMap.web3[key] = initWeb3Api(getCurrentProvider(this.networkMap[key]));
+      if (currentProvider) {
+        this.apiMap.dotSama[key] = initApi(key, currentProvider, this.networkMap[key].isEthereum);
+
+        if (this.networkMap[key].isEthereum && this.networkMap[key].isEthereum) {
+          this.apiMap.web3[key] = initWeb3Api(currentProvider);
+        }
       }
     }
 
@@ -1630,8 +1665,10 @@ export default class KoniState extends State {
 
     for (const [key, network] of Object.entries(this.networkMap)) {
       if (!network.active) {
-        if (networkKeys.includes(key)) {
-          this.apiMap.dotSama[key] = initApi(key, getCurrentProvider(this.networkMap[key]), this.networkMap[key].isEthereum);
+        const currentProvider = getCurrentProvider(this.networkMap[key]);
+
+        if (networkKeys.includes(key) && currentProvider) {
+          this.apiMap.dotSama[key] = initApi(key, currentProvider, this.networkMap[key].isEthereum);
           this.networkMap[key].active = true;
         }
       } else {
@@ -1705,7 +1742,11 @@ export default class KoniState extends State {
   }
 
   public refreshWeb3Api (key: string) {
-    this.apiMap.web3[key] = initWeb3Api(getCurrentProvider(this.networkMap[key]));
+    const currentProvider = getCurrentProvider(this.networkMap[key]);
+
+    if (currentProvider) {
+      this.apiMap.web3[key] = initWeb3Api(currentProvider);
+    }
   }
 
   public subscribeServiceInfo () {
@@ -1810,7 +1851,7 @@ export default class KoniState extends State {
   private combineHistories (oldItems: TransactionHistoryItemType[], newItems: TransactionHistoryItemType[]): TransactionHistoryItemType[] {
     const newHistories = newItems.filter((item) => !oldItems.some((old) => this.isSameHistory(old, item)));
 
-    return [...oldItems, ...newHistories].filter((his) => his.origin === 'app' || his.eventIdx).sort((a, b) => b.time - a.time);
+    return [...oldItems, ...newHistories].filter((his) => his.origin === 'app' || his.eventIdx);
   }
 
   public isSameHistory (oldItem: TransactionHistoryItemType, newItem: TransactionHistoryItemType): boolean {